- Rationalised the way that hardware definitions are imported.  Previously these were imported by the module hardware.py then inserted into the state machine definition modules by the hardware.initialise() function.  This made it very non-transparant from looking at the state machine description where the hardware object had come from.
In the revised code the hardware definition file is only imported by the state
machine definition and not by any of the framework modules.  This is a much more
sensible import structure and as a further advantage allows hardware objects
to be specified directly in the state machine definition without creaating a
seperate hardware definition file.

- The example state machines have been adapted to match this modification.