handle changed or missing variable error
After creating a custom variable dialog the user could potentially change the name of or delete a variable in the task file. The next time the task is uploaded the custom variable dialog.json file will potentially try to create a control widget for a variable that no longer exists.

Previously, this situation would throw an error in the terminal and then disconnect the board.

Now the task will still load, omitting the unfound variable,  and an error message for each missing control variable will appear in the log. Leftover (or newly named variables) will go into the "..." tab as usual.