Bugfixes for custom variable dialog (#35)
* fixed bug where leftover "add" button widgets were hiding beneath table

when a new row was added, a new "add" button was created, while the old add button still active but hidden underneath  the table. If the table columns were made skinnier, the extra "add" buttons were revealed, and were active/connected to the add row function. This commit fixes this bug, so only a single "add" button exists.

* handle changed or missing variable error

After creating a custom variable dialog the user could potentially change the name of or delete a variable in the task file. The next time the task is uploaded the custom variable dialog.json file will potentially try to create a control widget for a variable that no longer exists.

Previously, this situation would throw an error in the terminal and then disconnect the board.

Now the task will still load, omitting the unfound variable,  and an error message for each missing control variable will appear in the log. Leftover (or newly named variables) will go into the "..." tab as usual.